Phase3_Smart_Contract/requirements.txt
web3==5.31.3
eth-account==0.5.9
requests==2.31.0
python-dotenv==1.0.0
